Moved dependencies from ssh to http
Current git submodule dependencies are cloned through ssh.  While this form of git repository is valid, the move comes as a result of 1) adds an additional dependency on the build environment (openssh for example) and 2) the fact that these repositories are public anyway and authentication is not used to access their contents.